SweetAlert2 alert/warning/success messages are fully functional - Still need to make slight tweaks to EditPanel

+++ src/anime-db-handlers/handlers.js
@@ -1,9 +1,19 @@
+import { showDeleteConfirmAlert, showDeleteSuccessAlert } from "../swal-alerts/alerts"; 
+
 export const handleAnimeDeletion = async (animeMalId, shelfItems, setShelfItems, setAnimeDetails) => {
     try {
-        await window.dbFunctions.deleteAnime(animeMalId);
-        const updatedShelfItems = shelfItems.filter(anime => anime.mal_id !== animeMalId);
-        setShelfItems(updatedShelfItems);
-        setAnimeDetails(null);    
+        const confirmDeletion = await showDeleteConfirmAlert();
+        
+        if (!confirmDeletion) {
+            return; 
+            // Placeholder code, we need to ensure that it'll show a "Deletion Not Completed" alert in the future
+        } else {
+            await window.dbFunctions.deleteAnime(animeMalId);
+            const updatedShelfItems = shelfItems.filter(anime => anime.mal_id !== animeMalId);
+            setShelfItems(updatedShelfItems);
+            setAnimeDetails(null); 
+            showDeleteSuccessAlert();       
+        }
     } catch {
         console.log("Selected anime could not be deleted!");
     }
@@ -17,6 +27,11 @@ export const viewAnimeDetails = async (animeMalId, shelfItems, setAnimeDetails) 
 export const handleEditAnime = async (animeMalId, shelfItems, setEditingAnime) => {
     const animeToEdit = shelfItems.find(anime => anime.mal_id === animeMalId);
     setEditingAnime(animeToEdit);
+}
+
+export const handleAnimeAlreadyExists = async (animeMalId) => {
+    console.log("Hello World");
+    // Placeholder code for now
 }
 
 export const handleCloseDisplayPanel = async (setAnimeDetails, setAnime, setEditingAnime) => {
